requirements.txt
Django>=3.0,<4.0
psycopg2-binary>=2.8
djangorestframework==3.11.2
Pygments==2.7.4
Markdown==3.2.1
coreapi==2.3.3
dj-database-url==0.5.0
gunicorn==20.0.4
whitenoise==5.0.1
PyYAML==5.4
django-cors-headers>=3.0,<4.0
